requirements.txt
//functions-framework
azure-ai-formrecognizer
google-cloud-storage